backend/requirements.txt
# EchoNote Backend Dependencies
# FastAPI + SQLAlchemy + JWT

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Pydantic & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Audio Processing (可选，用于获取音频时长)
mutagen>=1.47.0

# HTTP Client
httpx>=0.25.0

# Environment
python-dotenv>=1.0.0
